feat: Refactor ToastProvider and Toast components
- Reorganize imports and fix import paths
- Update ToastProvider to use useEffect hook
- Update Toast component to use 10-second duration for auto-dismissal
- Create ToastsContainer component for rendering multiple toasts
- Add CSS styles for ToastsContainer component
- use dedicated portal created dynamically
