+++ src/components/ToastProvider/ToastProvider.tsx
@@ -1,9 +1,7 @@
-import { AnimatePresence, motion } from "framer-motion";
-import React, { createContext, useState } from "react";
-import { ToastType } from "../../types";
-import Toast from "../Toast";
-import { TOAST_DURATION, Variant } from "../../constants";
-import styles from "./ToastProvider.module.css";
+import React, { createContext, useEffect, useState } from "react";
+import { PositionType, ToastType } from "../../types";
+import { DEFAULT_TOAST_DURATION, Variant } from "../../constants";
+import ToastsContainer from "../ToastsContainer";
 
 export const ToastContext = createContext<{
   createToast: ({
@@ -35,14 +33,6 @@ export const ToastContext = createContext<{
   },
 });
 
-export type PositionType =
-  | "top-left"
-  | "top-right"
-  | "bottom-left"
-  | "bottom-right"
-  | "top-center"
-  | "bottom-center";
-
 function ToastProvider({
   children,
   position = "top-right",
@@ -54,7 +44,7 @@ function ToastProvider({
 
   function createToast({
     message,
-    duration = TOAST_DURATION,
+    duration = DEFAULT_TOAST_DURATION,
     variant,
   }: {
     message: string;
@@ -105,45 +95,18 @@ function ToastProvider({
     [setToasts]
   );
 
+  useEffect(() => {});
+
   return (
     <>
-      <ul
-        className={styles.wrapper}
-        style={{
-          top: position.startsWith("top") ? "1rem" : "auto",
-          bottom: position.startsWith("bottom") ? "1rem" : "auto",
-          left: position.endsWith("left")
-            ? "1rem"
-            : position.endsWith("center")
-            ? "50%"
-            : "auto",
-          right: position.endsWith("right") ? "1rem" : "auto",
-          transform: position.endsWith("center") ? "translateX(-50%)" : "none", // Corrected to translate to the left for centering
-        }}
-      >
-        <AnimatePresence mode="popLayout">
-          {toasts.map((toast) => (
-            <motion.li
-              layout
-              initial={{ scale: 0.8, opacity: 0 }}
-              animate={{ scale: 1, opacity: 1 }}
-              exit={{ scale: 0.8, opacity: 0 }}
-              transition={{ type: "spring" }}
-              className="box"
-              key={toast.id}
-            >
-              <Toast
-                duration={toast.duration}
-                variant={toast.variant}
-                id={toast.id}
-                onRemove={removeToast}
-              >
-                {toast.message}
-              </Toast>
-            </motion.li>
-          ))}
-        </AnimatePresence>
-      </ul>
+      {toasts.length > 0 && (
+        <ToastsContainer
+          position={position}
+          onRemove={removeToast}
+          toasts={toasts}
+        />
+      )}
+
       <ToastContext.Provider
         value={{ createToast, removeToast, removeAllToasts }}
       >
