+++ src/types.ts
@@ -6,3 +6,11 @@ export type ToastType = {
   message: string;
   variant: Variant;
 };
+
+export type PositionType =
+  | "top-left"
+  | "top-right"
+  | "bottom-left"
+  | "bottom-right"
+  | "top-center"
+  | "bottom-center";
